Uploads - Downloads
All uploaded files by the user saved in the database with base64
encoding at 'upload' table.
When the user requests a file to download, run a query to the database
in response brings the encoded file format.
then decoded and finally stored in a temporary folder of the user who
created that time.
Once the user exits from the system, or times out then the file and all
the temporary files it contains are deleted.